agent/requirements.txt
requests
mss
Pillow
psutil
python-dotenv
customtkinter
packaging
